Refactor(#44) : DTO 경로 리펙토링
