+++ src/main/java/com/keodam/keodam_backend/app/user/coffeeChatProfile/service/MenteeProfileService.java
@@ -3,8 +3,8 @@ package com.keodam.keodam_backend.app.user.coffeeChatProfile.service;
 import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.coffeeChatProfile.domain.Mentee;
 import com.keodam.keodam_backend.app.user.coffeeChatProfile.domain.MenteeHashtag;
-import com.keodam.keodam_backend.app.user.coffeeChatProfile.dto.MenteeRequestDto;
-import com.keodam.keodam_backend.app.user.coffeeChatProfile.dto.MenteeResponseDto;
+import com.keodam.keodam_backend.app.user.coffeeChatProfile.dto.request.MenteeRequestDto;
+import com.keodam.keodam_backend.app.user.coffeeChatProfile.dto.response.MenteeResponseDto;
 import com.keodam.keodam_backend.app.user.coffeeChatProfile.repository.MenteeHashtagRepository;
 import com.keodam.keodam_backend.app.user.coffeeChatProfile.repository.MenteeRepository;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
